Splitted CreateManualSymbolTask and the set price logic
