+++ GhostfolioSidekick/MarketDataMaintainer/CreateManualSymbolTask.cs
@@ -3,7 +3,6 @@ using GhostfolioSidekick.GhostfolioAPI;
 using GhostfolioSidekick.GhostfolioAPI.API.Mapper;
 using GhostfolioSidekick.Model;
 using GhostfolioSidekick.Model.Activities;
-using GhostfolioSidekick.Model.Activities.Types;
 using GhostfolioSidekick.Model.Symbols;
 using Microsoft.Extensions.Logging;
 
@@ -53,7 +52,6 @@ namespace GhostfolioSidekick.MarketDataMaintainer
 					}
 
 					await AddAndUpdateSymbol(symbolConfiguration, manualSymbolConfiguration);
-					await SetKnownPrices(symbolConfiguration, profiles, holdings);
 				}
 			}
 			catch (NotAuthorizedException)
@@ -63,92 +61,6 @@ namespace GhostfolioSidekick.MarketDataMaintainer
 			}
 
 			logger.LogDebug($"{nameof(CreateManualSymbolTask)} Done");
-		}
-
-		private async Task SetKnownPrices(SymbolConfiguration symbolConfiguration, List<SymbolProfile> profiles, List<Holding> holdings)
-		{
-			var mdi = profiles.SingleOrDefault(x =>
-				x.Symbol == symbolConfiguration.Symbol &&
-				x.DataSource == Datasource.MANUAL &&
-				x.AssetClass == EnumMapper.ParseAssetClass(symbolConfiguration.ManualSymbolConfiguration!.AssetClass) &&
-				x.AssetSubClass == EnumMapper.ParseAssetSubClass(symbolConfiguration.ManualSymbolConfiguration.AssetSubClass));
-			if (mdi == null || mdi.ActivitiesCount <= 0)
-			{
-				return;
-			}
-
-			var activitiesForSymbol = holdings
-				.Where(x =>
-					x.SymbolProfile?.Symbol == mdi.Symbol &&
-					x.SymbolProfile.DataSource == Datasource.MANUAL &&
-					x.SymbolProfile.AssetClass == EnumMapper.ParseAssetClass(symbolConfiguration.ManualSymbolConfiguration!.AssetClass) &&
-					x.SymbolProfile.AssetSubClass == EnumMapper.ParseAssetSubClass(symbolConfiguration.ManualSymbolConfiguration.AssetSubClass))
-				.SelectMany(x => x.Activities)
-				.OfType<BuySellActivity>()
-				.ToList();
-
-			if (!activitiesForSymbol.Any())
-			{
-				return;
-			}
-
-			var md = (await marketDataService.GetMarketData(mdi.Symbol, mdi.DataSource.ToString())).MarketData;
-			var sortedActivities = activitiesForSymbol
-				.Where(x => x.UnitPrice?.Amount != 0)
-				.GroupBy(x => x.Date.Date)
-				.Select(x => x
-					.OrderBy(x => x.TransactionId)
-					.ThenByDescending(x => x.UnitPrice?.Amount ?? 0)
-					.ThenByDescending(x => x.Quantity)
-					.First())
-				.OrderBy(x => x.Date)
-				.ToList();
-
-			for (var i = 0; i < sortedActivities.Count; i++)
-			{
-				var fromActivity = sortedActivities[i];
-				if (fromActivity?.UnitPrice == null)
-				{
-					continue;
-				}
-
-				BuySellActivity? toActivity = null;
-
-				if (i + 1 < sortedActivities.Count)
-				{
-					toActivity = sortedActivities[i + 1];
-				}
-
-				DateTime toDate = toActivity?.Date ?? DateTime.Today.AddDays(1);
-				for (var date = fromActivity.Date; date <= toDate; date = date.AddDays(1))
-				{
-					var a = (decimal)(date - fromActivity.Date).TotalDays;
-					var b = (decimal)(toDate - date).TotalDays;
-
-					var percentage = a / (a + b);
-					decimal amountFrom = fromActivity.UnitPrice!.Amount;
-					decimal amountTo = toActivity?.UnitPrice?.Amount ?? fromActivity.UnitPrice?.Amount ?? 0;
-					var expectedPrice = amountFrom + (percentage * (amountTo - amountFrom));
-
-					var price = md.SingleOrDefault(x => x.Date.Date == date.Date);
-
-					var diff = (price?.MarketPrice.Amount ?? 0) - expectedPrice;
-					if (Math.Abs(diff) >= Constants.Epsilon)
-					{
-						var scraperDefined = symbolConfiguration?.ManualSymbolConfiguration?.ScraperConfiguration != null;
-						var priceIsAvailable = (price?.MarketPrice.Amount ?? 0) != 0;
-						var isToday = date >= DateTime.Today;
-						var shouldSkip = scraperDefined && (priceIsAvailable || isToday);
-
-						if (shouldSkip)
-						{
-							continue;
-						}
-
-						await marketDataService.SetMarketPrice(mdi, new Money(fromActivity.UnitPrice!.Currency, expectedPrice), date);
-					}
-				}
-			}
 		}
 
 		private async Task AddAndUpdateSymbol(SymbolConfiguration symbolConfiguration, ManualSymbolConfiguration manualSymbolConfiguration)
